requirements.txt
# DJANGO
django>=1.7
djangorestframework